Adjusted title positioning and changed URLs
A couple of the urls pointed to areas of the repo that didn't exist, updated these with working URLs that matched the description.

Updated a couple of the titles to be above the images, this seems to look better. Provide your thoughts on whether this should be done for the remainder of the images.